chore: Move shared schemas into a lightweight runner export, eliminating circular dependencies. In service of #2245
