+++ packages/charm/src/index.ts
@@ -1,15 +1,4 @@
-export {
-  charmId,
-  charmListSchema,
-  CharmManager,
-  favoriteListSchema,
-  getRecipeIdFromCharm,
-  type NameSchema,
-  nameSchema,
-  processSchema,
-  type UISchema,
-  uiSchema,
-} from "./manager.ts";
+export { charmId, CharmManager, getRecipeIdFromCharm } from "./manager.ts";
 export {
   addFavorite,
   getHomeFavorites,
