+++ packages/charm/src/manager.ts
@@ -9,21 +9,26 @@ import {
   JSONSchema,
   type MemorySpace,
   Module,
-  NAME,
   parseLink,
   Recipe,
   Runtime,
   type Schema,
   type SpaceCellContents,
   TYPE,
-  UI,
   URI,
 } from "@commontools/runner";
 import * as favorites from "./favorites.ts";
 import { ALL_CHARMS_ID } from "../../runner/src/builtins/well-known.ts";
-import { vdomSchema } from "@commontools/html";
 import { type Session } from "@commontools/identity";
 import { isObject, isRecord } from "@commontools/utils/types";
+import {
+  charmListSchema,
+  charmSourceCellSchema,
+  FavoriteList,
+  NameSchema,
+  nameSchema,
+  processSchema,
+} from "@commontools/runner/schemas";
 
 /**
  * Extracts the ID from a charm.
@@ -36,80 +41,6 @@ export function charmId(charm: Cell<unknown>): string | undefined {
   const idValue = id["/"];
   return typeof idValue === "string" ? idValue : undefined;
 }
-
-export const nameSchema = {
-  type: "object",
-  properties: { [NAME]: { type: "string" } },
-  required: [NAME],
-} as const satisfies JSONSchema;
-
-export type NameSchema = Schema<typeof nameSchema>;
-
-export const uiSchema = {
-  type: "object",
-  properties: { [UI]: vdomSchema },
-  required: [UI],
-} as const satisfies JSONSchema;
-
-export type UISchema = Schema<typeof uiSchema>;
-
-export const charmListSchema = {
-  type: "array",
-  items: { not: true, asCell: true },
-} as const satisfies JSONSchema;
-
-export const charmLineageSchema = {
-  type: "object",
-  properties: {
-    charm: { not: true, asCell: true },
-    relation: { type: "string" },
-    timestamp: { type: "number" },
-  },
-  required: ["charm", "relation", "timestamp"],
-} as const satisfies JSONSchema;
-export type CharmLineage = Schema<typeof charmLineageSchema>;
-
-export const favoriteEntrySchema = {
-  type: "object",
-  properties: {
-    cell: { not: true, asCell: true },
-    tag: { type: "string", default: "" },
-  },
-  required: ["cell"],
-} as const satisfies JSONSchema;
-
-export type FavoriteEntry = Schema<typeof favoriteEntrySchema>;
-
-export const favoriteListSchema = {
-  type: "array",
-  items: favoriteEntrySchema,
-} as const satisfies JSONSchema;
-
-export type FavoriteList = Schema<typeof favoriteListSchema>;
-
-export const charmSourceCellSchema = {
-  type: "object",
-  properties: {
-    [TYPE]: { type: "string" },
-    spell: { type: "object" },
-    lineage: {
-      type: "array",
-      items: charmLineageSchema,
-      default: [],
-    },
-    llmRequestId: { type: "string" },
-  },
-} as const satisfies JSONSchema;
-
-export const processSchema = {
-  type: "object",
-  properties: {
-    argument: { type: "object" },
-    [TYPE]: { type: "string" },
-    spell: { type: "object" },
-  },
-  required: [TYPE],
-} as const satisfies JSONSchema;
 
 /**
  * Filters an array of charms by removing any that match the target cell
