+++ packages/html/src/index.ts
@@ -4,7 +4,6 @@ export {
   setEventSanitizer,
   setNodeSanitizer,
   type SetPropHandler,
-  vdomSchema,
 } from "./render.ts";
 export { isVNode, type VNode } from "./jsx.ts";
 export { h } from "./h.ts";
