+++ packages/runner/src/shared.ts
@@ -1,0 +1,29 @@
+/**
+ * /!\ Shared between client and runtime threads.
+ * /!\ Take care in only importing lightweight types,
+ * /!\ interfaces and utilities.
+ */
+
+export {
+  isLegacyAlias,
+  isSigilLink,
+  type NormalizedFullLink,
+  parseLLMFriendlyLink,
+} from "./link-types.ts";
+export { LINK_V1_TAG, type SigilLink, type URI } from "./sigil-types.ts";
+export {
+  ID,
+  type JSONSchema,
+  type JSONValue,
+  NAME,
+  type Schema,
+  TYPE,
+  UI,
+} from "./builder/types.ts";
+export { effect } from "./reactivity.ts";
+export { type Cancel, useCancelGroup } from "./cancel.ts";
+export type {
+  RuntimeTelemetry,
+  RuntimeTelemetryEvent,
+  RuntimeTelemetryMarkerResult,
+} from "./telemetry.ts";
